:truck: move common/ThemeProvider into components/ThemeProvider
