+++ apps/github-search/src/pages/_app.tsx
@@ -2,7 +2,7 @@ import React, { FC } from "react";
 import type { AppProps } from "next/app";
 import { ApolloProvider } from "@apollo/client";
 
-import ThemeProvider from "@/common/ThemeProvider";
+import ThemeProvider from "@/components/ThemeProvider";
 import client from "@/utils/apollo-client";
 
 const App: FC<AppProps> = ({ Component, pageProps }) => (
