+++ apps/github-search/test/test-utils.ts
@@ -1,6 +1,6 @@
 import { ReactElement } from "react";
 import { render, RenderOptions, RenderResult } from "@testing-library/react";
-import ThemeProvider from "@/common/ThemeProvider";
+import ThemeProvider from "@/components/ThemeProvider";
 
 const customRender = (ui: ReactElement, options?: Omit<RenderOptions, "quries">): RenderResult =>
   render(ui, { wrapper: ThemeProvider, ...options });
